+++ atlasdb-cassandra-integration-tests/src/test/java/com/palantir/atlasdb/keyvalue/cassandra/SchemaMutationLockTablesIntegrationTest.java
@@ -17,15 +17,14 @@
 package com.palantir.atlasdb.keyvalue.cassandra;
 
 import static org.hamcrest.Matchers.empty;
-import static org.hamcrest.Matchers.hasItem;
+import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.hasSize;
 import static org.hamcrest.Matchers.is;
-import static org.hamcrest.Matchers.startsWith;
 import static org.junit.Assert.assertThat;
+import static org.junit.Assert.fail;
 
-import java.util.ArrayList;
 import java.util.Collections;
-import java.util.List;
+import java.util.HashSet;
 import java.util.Set;
 import java.util.UUID;
 import java.util.concurrent.Callable;
@@ -33,6 +32,12 @@ import java.util.concurrent.CyclicBarrier;
 import java.util.function.IntConsumer;
 import java.util.stream.IntStream;
 
+import org.apache.cassandra.db.marshal.UUIDType;
+import org.apache.cassandra.thrift.Column;
+import org.apache.cassandra.thrift.Compression;
+import org.apache.cassandra.thrift.ConsistencyLevel;
+import org.apache.cassandra.thrift.CqlResult;
+import org.apache.cassandra.thrift.InvalidRequestException;
 import org.apache.thrift.TException;
 import org.junit.Before;
 import org.junit.ClassRule;
@@ -44,6 +49,7 @@ import com.palantir.atlasdb.cassandra.ImmutableCassandraKeyValueServiceConfig;
 import com.palantir.atlasdb.containers.CassandraContainer;
 import com.palantir.atlasdb.containers.Containers;
 import com.palantir.atlasdb.keyvalue.api.TableReference;
+import com.palantir.logsafe.SafeArg;
 
 public class SchemaMutationLockTablesIntegrationTest {
     @ClassRule
@@ -77,10 +83,17 @@ public class SchemaMutationLockTablesIntegrationTest {
     }
 
     @Test
-    public void multipleLockTablesExistAfterCreation() throws Exception {
+    public void secondTableCreationFailsAndOnlyOneTableExistsAfterwards() throws Exception {
         lockTables.createLockTable();
-        lockTables.createLockTable();
-        assertThat(lockTables.getAllLockTables(), hasSize(2));
+
+        try {
+            lockTables.createLockTable();
+            fail("Should have thrown an InvalidRequestException!");
+        } catch (InvalidRequestException ire) {
+            // expected
+        }
+
+        assertThat(lockTables.getAllLockTables(), hasSize(1));
     }
 
     @Test
@@ -91,33 +104,59 @@ public class SchemaMutationLockTablesIntegrationTest {
     }
 
     @Test
-    public void shouldCreateLockTablesStartingWithCorrectPrefix() throws TException {
+    public void shouldCreateLockTablesWithCorrectName() throws TException {
         lockTables.createLockTable();
 
-        assertThat(Iterables.getOnlyElement(lockTables.getAllLockTables()).getTablename(), startsWith("_locks_"));
+        assertThat(Iterables.getOnlyElement(lockTables.getAllLockTables()).getTablename(), equalTo("_locks"));
     }
 
     @Test
-    public void whenTablesAreCreatedConcurrentlyAtLeastOneThreadShouldSeeBothTables() {
+    public void shouldCreateLockTableWithFixedCfId() throws Exception {
+        lockTables.createLockTable();
+
+        TableReference lockTable = Iterables.getOnlyElement(lockTables.getAllLockTables());
+        String query = "SELECT cf_id FROM system.schema_columnfamilies WHERE keyspace_name = '%s' "
+                + "AND columnfamily_name = '%s'";
+        CqlQuery cqlQuery = new CqlQuery(query, SafeArg.of("keyspace", config.getKeyspaceOrThrow()),
+                SafeArg.of("tableName", lockTable.getTablename()));
+        byte[] storedBytes = clientPool.run(client -> {
+            CqlResult cqlResult = client.execute_cql3_query(cqlQuery, Compression.NONE, ConsistencyLevel.QUORUM);
+            Column column = cqlResult.getRows().stream().findFirst().get().getColumns().get(0);
+            return column.getValue();
+        });
+
+        String fullTableName = config.getKeyspaceOrThrow() + "." + lockTable.getTablename();
+        UUID expectedUuid = UUID.nameUUIDFromBytes(fullTableName.getBytes());
+        byte[] expectedBytes = UUIDType.instance.fromString(expectedUuid.toString()).array();
+
+        assertThat("Created a lock table with the wrong ID!",
+                storedBytes, equalTo(expectedBytes));
+
+    }
+
+    @Test
+    public void whenTablesAreCreatedConcurrentlyThereIsStillOnlyOneTable() {
         CyclicBarrier barrier = new CyclicBarrier(2);
 
-        List<Set<TableReference>> lockTablesSeen = Collections.synchronizedList(new ArrayList<>());
+        Set<TableReference> lockTablesSeen = Collections.synchronizedSet(new HashSet<>());
 
         IntStream.range(0, 2).parallel()
                 .forEach(ignoringExceptions(() -> {
                     barrier.await();
                     lockTables.createLockTable();
-                    lockTablesSeen.add(lockTables.getAllLockTables());
+                    lockTablesSeen.addAll(lockTables.getAllLockTables());
                     return null;
                 }));
 
-        assertThat("Only one table was seen by both creation threads", lockTablesSeen, hasItem(hasSize(2)));
+        assertThat("Only one table should have been seen by both creation threads", lockTablesSeen, hasSize(1));
     }
 
     private IntConsumer ignoringExceptions(Callable function) {
         return (iterationCount) -> {
             try {
                 function.call();
+            } catch (InvalidRequestException ire) {
+                // ignore - expected if tables aren't quite created at the same time.
             } catch (Exception e) {
                 throw new RuntimeException(e);
             }