+++ docs/source/release_notes/release-notes.rst
@@ -50,8 +50,10 @@ develop
     *    - Type
          - Change
 
-    *    -
-         -
+    *    - |fixed|
+         - The ``_locks`` table is now created with a deterministic column family ID.
+         This means that multi-node installations will no longer create multiple ``_locks_<uuid>`` tables on first start-up.
+           (`Pull Request <https://github.com/palantir/atlasdb/pull/3088>`__)
 
 =======
 v0.80.0
@@ -65,7 +67,7 @@ v0.80.0
 
     *    - Type
          - Change
-    
+
     *    - |fixed| |devbreak|
          - Centralize how ``PersistentLockManager`` is created in a dagger context.
            Also, removed the old constructor for ``CellsSweeper``.